// apps/api/src/common/config/app/app-config.type.ts
export type AppConfig = {
    apiPrefix: string;
    gmailPass: string;
    gmailUser: string;
    headerLanguage: string;
    isProduction: boolean;
    nodeEnv: string;
    origin: string;
    frontendOrigin: string;
    port: number;
    resendApiKey: string;
    resendFrom: string;
    workingDirectory: string;
    version: string;
};